+++ backend/services/notice/service/__init__.py
@@ -1,0 +1,3 @@
+from .me import *
+from .default import *
+from .base import *
